+++ pkgs/native_assets_cli/test/model/build_output_test.dart
@@ -109,4 +109,54 @@ version: ${BuildOutput.version}''';
       );
     });
   }
+
+  test('format exception', () {
+    expect(
+      () => BuildOutput.fromYamlString('''timestamp: 2022-11-10 13:25:01.000
+assets:
+  - name: foo
+    link_mode: dynamic
+    path:
+      path_type:
+        some: map
+      uri: path/to/libfoo.so
+    target: android_x64
+dependencies: []
+metadata:
+  key: value
+version: ${BuildOutput.version}'''),
+      throwsFormatException,
+    );
+    expect(
+      () => BuildOutput.fromYamlString('''timestamp: 2022-11-10 13:25:01.000
+assets:
+  - name: foo
+    link_mode: dynamic
+    path:
+      path_type: absolute
+      uri: path/to/libfoo.so
+    target: android_x64
+dependencies:
+  1: foo
+metadata:
+  key: value
+version: ${BuildOutput.version}'''),
+      throwsFormatException,
+    );
+    expect(
+      () => BuildOutput.fromYamlString('''timestamp: 2022-11-10 13:25:01.000
+assets:
+  - name: foo
+    link_mode: dynamic
+    path:
+      path_type: absolute
+      uri: path/to/libfoo.so
+    target: android_x64
+dependencies: []
+metadata:
+  123: value
+version: ${BuildOutput.version}'''),
+      throwsFormatException,
+    );
+  });
 }