+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added topics.
 - Fixed metadata example.
+- Throws `FormatException`s instead of `TypeError`s when failing to parse Yaml
+  ([#109](https://github.com/dart-lang/native/issues/109)).
 
 ## 0.1.0