+++ pkgs/native_assets_cli/CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.1.1-wip
+## 0.2.0
 
+- *Breaking change* Rename `Asset.name` to `Asset.id`
+  ([#100](https://github.com/dart-lang/native/issues/100)).
 - Added topics.
 - Fixed metadata example.
 - Throws `FormatException`s instead of `TypeError`s when failing to parse Yaml